// models/CreateVariantRequest.ts
/**
 * Flagr
 * Flagr is a feature flagging, A/B testing and dynamic configuration microservice. The base path for all the APIs is \"/api/v1\". 
 *
 * OpenAPI spec version: 1.1.13
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateVariantRequest {
    'key': string;
    'attachment'?: any;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "key",
            "baseName": "key",
            "type": "string",
            "format": ""
        },
        {
            "name": "attachment",
            "baseName": "attachment",
            "type": "any",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateVariantRequest.attributeTypeMap;
    }
    
    public constructor() {
    }
}