// models/Distribution.ts
/**
 * Flagr
 * Flagr is a feature flagging, A/B testing and dynamic configuration microservice. The base path for all the APIs is \"/api/v1\". 
 *
 * OpenAPI spec version: 1.1.13
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Distribution {
    'id'?: number;
    'percent': number;
    'variantKey': string;
    'variantID': number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "percent",
            "baseName": "percent",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "variantKey",
            "baseName": "variantKey",
            "type": "string",
            "format": ""
        },
        {
            "name": "variantID",
            "baseName": "variantID",
            "type": "number",
            "format": "int64"
        }    ];

    static getAttributeTypeMap() {
        return Distribution.attributeTypeMap;
    }
    
    public constructor() {
    }
}