// models/EvalContext.ts
/**
 * Flagr
 * Flagr is a feature flagging, A/B testing and dynamic configuration microservice. The base path for all the APIs is \"/api/v1\". 
 *
 * OpenAPI spec version: 1.1.13
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class EvalContext {
    /**
    * entityID is used to deterministically at random to evaluate the flag result. If it's empty, flagr will randomly generate one.
    */
    'entityID'?: string;
    'entityType'?: string;
    'entityContext'?: any;
    'enableDebug'?: boolean;
    /**
    * flagID
    */
    'flagID'?: number;
    /**
    * flagKey. flagID or flagKey will resolve to the same flag. Either works.
    */
    'flagKey'?: string;
    /**
    * flagTags. flagTags looks up flags by tag. Either works.
    */
    'flagTags'?: Array<string>;
    /**
    * determine how flagTags is used to filter flags to be evaluated. OR extends the evaluation to those which contains at least one of the provided flagTags or AND limit the evaluation to those which contains all the flagTags.
    */
    'flagTagsOperator'?: EvalContextFlagTagsOperatorEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "entityID",
            "baseName": "entityID",
            "type": "string",
            "format": ""
        },
        {
            "name": "entityType",
            "baseName": "entityType",
            "type": "string",
            "format": ""
        },
        {
            "name": "entityContext",
            "baseName": "entityContext",
            "type": "any",
            "format": ""
        },
        {
            "name": "enableDebug",
            "baseName": "enableDebug",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "flagID",
            "baseName": "flagID",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "flagKey",
            "baseName": "flagKey",
            "type": "string",
            "format": ""
        },
        {
            "name": "flagTags",
            "baseName": "flagTags",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "flagTagsOperator",
            "baseName": "flagTagsOperator",
            "type": "EvalContextFlagTagsOperatorEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return EvalContext.attributeTypeMap;
    }
    
    public constructor() {
    }
}


export type EvalContextFlagTagsOperatorEnum = "ANY" | "ALL" ;