// models/EvalDebugLog.ts
/**
 * Flagr
 * Flagr is a feature flagging, A/B testing and dynamic configuration microservice. The base path for all the APIs is \"/api/v1\". 
 *
 * OpenAPI spec version: 1.1.13
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { SegmentDebugLog } from './SegmentDebugLog';
import { HttpFile } from '../http/http';

export class EvalDebugLog {
    'segmentDebugLogs'?: Array<SegmentDebugLog>;
    'msg'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "segmentDebugLogs",
            "baseName": "segmentDebugLogs",
            "type": "Array<SegmentDebugLog>",
            "format": ""
        },
        {
            "name": "msg",
            "baseName": "msg",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return EvalDebugLog.attributeTypeMap;
    }
    
    public constructor() {
    }
}