// models/EvaluationBatchRequest.ts
/**
 * Flagr
 * Flagr is a feature flagging, A/B testing and dynamic configuration microservice. The base path for all the APIs is \"/api/v1\". 
 *
 * OpenAPI spec version: 1.1.13
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { EvaluationEntity } from './EvaluationEntity';
import { HttpFile } from '../http/http';

export class EvaluationBatchRequest {
    'entities': Array<EvaluationEntity>;
    'enableDebug'?: boolean;
    /**
    * flagIDs
    */
    'flagIDs'?: Array<number>;
    /**
    * flagKeys. Either flagIDs, flagKeys or flagTags works. If pass in multiples, Flagr may return duplicate results.
    */
    'flagKeys'?: Array<string>;
    /**
    * flagTags. Either flagIDs, flagKeys or flagTags works. If pass in multiples, Flagr may return duplicate results.
    */
    'flagTags'?: Array<string>;
    /**
    * determine how flagTags is used to filter flags to be evaluated. OR extends the evaluation to those which contains at least one of the provided flagTags or AND limit the evaluation to those which contains all the flagTags.
    */
    'flagTagsOperator'?: EvaluationBatchRequestFlagTagsOperatorEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "entities",
            "baseName": "entities",
            "type": "Array<EvaluationEntity>",
            "format": ""
        },
        {
            "name": "enableDebug",
            "baseName": "enableDebug",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "flagIDs",
            "baseName": "flagIDs",
            "type": "Array<number>",
            "format": "int64"
        },
        {
            "name": "flagKeys",
            "baseName": "flagKeys",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "flagTags",
            "baseName": "flagTags",
            "type": "Array<string>",
            "format": ""
        },
        {
            "name": "flagTagsOperator",
            "baseName": "flagTagsOperator",
            "type": "EvaluationBatchRequestFlagTagsOperatorEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return EvaluationBatchRequest.attributeTypeMap;
    }
    
    public constructor() {
    }
}


export type EvaluationBatchRequestFlagTagsOperatorEnum = "ANY" | "ALL" ;