// models/EvaluationBatchResponse.ts
/**
 * Flagr
 * Flagr is a feature flagging, A/B testing and dynamic configuration microservice. The base path for all the APIs is \"/api/v1\". 
 *
 * OpenAPI spec version: 1.1.13
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { EvalResult } from './EvalResult';
import { HttpFile } from '../http/http';

export class EvaluationBatchResponse {
    'evaluationResults': Array<EvalResult>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "evaluationResults",
            "baseName": "evaluationResults",
            "type": "Array<EvalResult>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return EvaluationBatchResponse.attributeTypeMap;
    }
    
    public constructor() {
    }
}